requirements.txt
flask==3.0.0
google-generativeai==0.3.0
torch==2.1.0
transformers==4.36.0